feat(list-route): 테스팅 코드 삭제
Resolve: #148
